+++ src/main/java/community/independe/api/dtos/post/PostsResponse.java
@@ -16,7 +16,7 @@ public class PostsResponse {
     private String title;
     private LocalDateTime createdDate;
     private Integer views; // 조회수
-    private Integer recommendCount; // 추천수
+    private Long recommendCount; // 추천수
     private Long commentCount; // 댓글수
     private boolean isPicture;
 }
